Call me a weirdo, but
I was creating my /about.html for my own webpage, so
I decided to add an image of me and my ThinkPad,
and then I thought to use your images' width and height
that you use in your /about.html. I decided to use w3m for that,
so I entered into the source of your /about.html
(It was a mess, HTML is a mess in general),
and found a stray double quote in pearl-funky.png image tag,
I thought it was some site generator error (what else could it be?)
I went to your daringcuteseal.github.io repo to check,
and I found that you wrote it in plain HTML?
(who write their about page in HTML?)
And the stray quote was there, in plain sight,
I thought to my self "Oops, I got too far. AGAIN!".
Then I decided to fix it and open a pull request.
I already gone too far, so why not?